requirements.txt
PyYAML==3.12
Pillow==6.2.2
